requirements.txt
pyTelegramBotAPI==4.12.0
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
requests==2.31.0
